feat : 유저 업데이트 api 개발

+++ src/main/java/com/youngjun/sns/user/dto/response/UserResponse.java
@@ -8,7 +8,7 @@ public class UserResponse {
     private Long id;
     private String email;
     private String phoneNumber;
-    private Long age;
+    private int age;
 
     public UserResponse(User user){
         this.id = user.getId();
